consolidate gpt tools
