+++ apps/wizard/scripts/news.py
@@ -4,9 +4,9 @@ from typing import Tuple
 import requests
 from structlog import get_logger
 
+from apps.utils.gpt import GPTQuery, OpenAIWrapper, get_cost_and_tokens
 from apps.wizard.app_pages.expert.prompts import SYSTEM_PROMPT_GENERIC
 from apps.wizard.utils.db import WizardDB
-from apps.wizard.utils.gpt import GPTQuery, OpenAIWrapper, get_cost_and_tokens
 
 # Logger
 log = get_logger()
